Move controller-specific classes to a dedicated package
